refactor : 디렉토리 명 변경
