+++ src/main/java/racingcar/validator/ErrorCode.java
@@ -1,4 +1,4 @@
-package racingcar.validation;
+package racingcar.validator;
 
 public enum ErrorCode {
     // common
