+++ src/test/java/racingcar/validator/CarNamesValidatorTest.java
@@ -1,4 +1,4 @@
-package racingcar.validation;
+package racingcar.validator;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
